Make sure hyperlinks across different solutions work fine.
We had some logic where we wouldn't generate a hyperlink to a symbol if it wasn't located in the current solution. This resulted in bug #6.

The way to fix it is first load all solutions, collect all assembly names from all solutions and then use that global list to determine whether a project was included in the index.

This solution doesn't depend on the order in which projects or solutions are included in the index.