Fix loading files with commas in the name
A file's URL should only contain a comma if it directly precedes a
line number. In the cases where a file's name actually contains a
comma, it should be encoded as %2C so the URL can still be split on
comma. This change prevents the encoded comma from being decoded
prematurely, resulting in an incorrect `potentialFile`. Furthermore,
it's completely unnecessary as `decodeURIComponent` on line 104 will
handle the decoding.