scripts.js: handle filenames with commas
The "anchorSplitChar" is used to separate the elements of the URL between the filename, the hash or line number if present, and the type of search (namespaces, references, or neither) if present. Currently this character is the comma. Unfortunately comma is a permissible character in filenames as well; in fact, a comma exists in a filename in the test solution (https://github.com/KirillOsenkov/SourceBrowser/blob/master/TestCode/Project2/-._%7E!%24'()%2B%2C%3D%40.cs), so splitting on comma alone is not safe as it results in "filenames" that are missing any text after the comma.

Three solutions come to mind:
- Change the split character to something else, for example a URL-safe character that is not usually allowed in filenames, like `*`. This would require finding all the relevant places literal commas are hard-coded and seemed error-prone, and would also (debatably) make the URL's less human-readable.
- Change all the places that generate the URL's to encode commas in filenames. After some investigation of this option, finding and fixing all of those seemed more complex than the third alternative.
- Change the places that parse the URL's to use the context around the comma to understand if it's an element-splitting comma or part of a filename. That's what I've done in this commit.